DPBOT_win/app/requirements.txt
tencentcloud_sdk_python_common
tencentcloud_sdk_python_hunyuan
fastapi
feedparser==6.0.11
lxml==5.2.2
requests==2.32.3
urllib3==2.2.3
websockets==11.0.3
meme_generator==0.0.19.1
Pillow==10.0.0
pydub==0.25.1
colorama==0.4.6
loguru==0.7.0
lz4
pyyaml==6.0
python-dotenv==1.0.0
schedule==1.2.2
qrcode==7.4.2